requirements.txt
Flask
google-cloud-datastore
python-jose
flask-cors
six
python-dotenv
requests
authlib
pyopenssl